:card_file_box: Switch database connection to PostgreSQL
Continuing to use Amazon RDS would have resulted in security issues and
a lack of ability to use it locally. Additionally, there weren't any
migrations or MySQL specific code. So @gannondetroit and I decided to
switch to PostgreSQL, which can be used as a free heroku addon, and
keeps our backend infrastructure together.