:bug: fixed the stripe api functionality
Stripe fully works and we have the ability to subscribe, upgrade our subscription and cancel our subscription. I replaced all the keys with the new ones I made on the Stripe dashboard. I did something weird when jumping between branches so I deleted and re-copied the file and there does not seem to be any issue.