:recycle: Move database to TrainingBot v2 data model
After working to refactor the application, we came to the conclusion
that the application would not scale and adding our mvp features was not
going to be possible. We designed a new datamodel to provide us with
scalability and stability and allow us to pass the application on to
the next team without issue.