:sparkles: Added for_team_member column for notifs
Notifications now has for_team_member boolean which allows for much simpler filtering on the front end as well as making it immediately clear when viewing the notif whether it was specifically for the "assigned" team member or if it was a copy meant for a mentor or manager. Also added fields to be returned for each notification and Joi validation.